requirements.txt
agno
openai
python-dotenv
pydantic
rich
fastapi
uvicorn
